requirements.txt
streamlit==1.31.0
pandas
plotly.express
openpyxl
lxml
requests
beautifulsoup4
